requirements.txt
amqp==2.6.1
asgiref==3.2.10
attrs==20.2.0
billiard==3.6.3.0
cached-property==1.5.2
cachetools==4.1.1
celery==4.4.7
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.1.1
decorator==4.4.2
defusedxml==0.7.0rc1
Django==3.1.1
django-autotranslate==1.1.1
django-braces==1.14.0
django-ckeditor==6.0.0
django-filter==2.3.0
django-fsm==2.7.0
django-js-asset==1.2.2
django-modeltranslation==0.15.2
django-mptt==0.11.0
django-multiurl==1.4.0
django-nested-admin==3.3.2
django-oauth-toolkit==1.3.2
django-recaptcha3==0.2.0
django-rest-framework-social-oauth2==1.1.0
django-rest-swagger==2.2.0
django-simple-history==2.11.0
django-solo==1.1.3
django-uuidfield==0.5.0
django-uuslug==1.2.0
djangorestframework==3.11.1
docopt==0.6.2
easy-thumbnails==2.7
fcm-django==0.3.4
flake8==3.8.3
future==0.18.2
futures==3.1.1
google-api-core==1.22.2
google-auth==1.21.2
google-cloud-core==1.4.1
google-cloud-translate==3.0.1
googleapis-common-protos==1.52.0
goslate==1.5.1
grpcio==1.32.0
gunicorn==20.0.4
idna==2.10
importlib-metadata==2.0.0
itypes==1.2.0
Jinja2==2.11.2
jsonschema==3.2.0
kombu==4.6.11
libcst==0.3.10
Markdown==3.2.2
MarkupSafe==1.1.1
mccabe==0.6.1
model-mommy==2.0.0
mypy-extensions==0.4.3
oauthlib==3.1.0
openapi-codec==1.3.2
packaging==20.4
Pillow==7.2.0
pip-review==1.1.0
polib==1.1.0
proto-plus==1.9.1
protobuf==3.13.0
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pyfcm==1.4.7
pyflakes==2.2.0
PyJWT==1.7.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.14.0
python-monkey-business==1.0.0
python-slugify==4.0.1
python-telegram-bot==12.8
python3-openid==3.2.0
pytils==0.3
pytz==2020.1
PyYAML==5.3.1
redis==3.5.3
requests==2.24.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rsa==4.6
simplejson==3.17.2
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.3.1
text-unidecode==1.3
texttable==1.6.3
tornado==6.0.4
transliterate==1.10.2
typing-extensions==3.7.4.3
typing-inspect==0.6.0
Unidecode==1.1.1
uritemplate==3.0.1
urllib3==1.25.10
viberbot==1.0.11
websocket-client==0.57.0
zipp==3.2.0
nats-python==0.7.0
asyncio-nats-client==0.11.4
drf-yasg==1.17.1